Moved the pluginConfiguration to azure package and renamed it
