+++ git-plugin/src/main/kotlin/io/rabobank/ret/git/plugin/provider/GitProviderSelector.kt
@@ -1,6 +1,6 @@
 package io.rabobank.ret.git.plugin.provider
 
-import io.rabobank.ret.git.plugin.config.PluginConfig
+import io.rabobank.ret.git.plugin.provider.azure.AzureDevopsPluginConfig
 import io.rabobank.ret.git.plugin.provider.azure.AzureDevopsClient
 import io.rabobank.ret.git.plugin.provider.azure.AzureDevopsProvider
 import io.rabobank.ret.git.plugin.provider.azure.AzureDevopsUrlFactory
@@ -8,7 +8,7 @@ import jakarta.enterprise.context.ApplicationScoped
 import jakarta.enterprise.inject.Produces
 
 @ApplicationScoped
-class GitProviderSelector(private val azureDevopsClient: AzureDevopsClient, private val pluginConfig: PluginConfig, private val azureDevopsUrlFactory: AzureDevopsUrlFactory) {
+class GitProviderSelector(private val azureDevopsClient: AzureDevopsClient, private val pluginConfig: AzureDevopsPluginConfig, private val azureDevopsUrlFactory: AzureDevopsUrlFactory) {
 
     @Produces
     fun gitProvider() = AzureDevopsProvider(azureDevopsClient, pluginConfig, azureDevopsUrlFactory)
