+++ git-plugin/src/main/kotlin/io/rabobank/ret/git/plugin/provider/azure/AzureDevopsConfiguration.kt
@@ -1,6 +1,5 @@
 package io.rabobank.ret.git.plugin.provider.azure
 
-import io.rabobank.ret.git.plugin.config.PluginConfig
 import jakarta.enterprise.context.ApplicationScoped
 import jakarta.enterprise.inject.Produces
 import jakarta.ws.rs.core.UriBuilder
@@ -10,7 +9,7 @@ import java.net.URI
 
 @ApplicationScoped
 class AzureDevopsConfiguration(
-    private val pluginConfig: PluginConfig,
+    private val pluginConfig: AzureDevopsPluginConfig,
     @ConfigProperty(name = "azure.devops.baseUrl") private val azureDevopsBaseUrl: String,
 ) {
